### 2025-06-18 #8432 @editorialbot
[PRE REVIEW]: NADI -- Network Analysis and Data Integration with a Domain Specific Programming Language
**Submitting author:** @atreyagaurav (<a href="http://orcid.org/0000-0002-0234-2165">Gaurav Atreya</a>)
**Repository:** https://github.com/Nadi-System/nadi-system
**Branch with paper.md** (empty if default branch):
**Version:** v0.7.0
