### 2025-06-19 #8432 @Atreyagaurav
Hi,

Thank you for starting the process. I'll go over the instructions and make changes.

A quick question though, I saw it says the word limit is 250-1000 and my paper is around 1700, should I shorten it?
