### 2025-06-19 41e53c1
shorten the paper
