+++ paper.md
@@ -23,11 +23,6 @@ date: 20 May 2025
 bibliography: references.bib
 ---
 
-<!-- author: -->
-<!--   - Gaurav Atreya -->
-<!--   - Todd Steissberg -->
-<!--   - Patrick A. Ray -->
-
 # Summary
 We present Network Analysis and Data Integration (NADI) System, a
 developing software framework designed to facilitate river data
@@ -35,55 +30,58 @@ analysis and visualization. NADI comes with a Domain Specific
 Programming Language (DSL) that has an intuitive syntax for network
 metadata analysis as well as a generalized plugin system to run
 user-defined functions on each node or the whole network. Plugins
-provide seamless integration with GIS softwares and and other
-programming languages. Our approach enables users to easily run
-network based algorithms, check for inconsistencies in data, easily
-visualize network metadata, etc. ultimately enhancing the accuracy and
-comprehensiveness of hydrological models compared to using only
-specialized software or general purpose programming languages.
+provide seamless integration with other softwares and
+programming languages.
 
 # Statement of need
-Hydrologic modeling, which involves integrating diverse data to simulate complex processes [@singhMathematicalModelingWatershed2002; @clarkImprovingRepresentationHydrologic2015; @loucksWaterResourceSystems2017], is hindered by manual and time-consuming data organization tasks, resulting on time drain, disengagement, information overload and human errors [@brunsTediousWorkDeveloping2024, @readStateScienceEvolving2021]
-highlighting the need for intelligent computational assistance to reduce the workload and improve the efficiency and reproducibility of research in this field [@rosenbergNextFrontierMaking2020].
+Hydrologic modeling, which involves integrating diverse data to
+simulate complex processes [@singhMathematicalModelingWatershed2002;
+@clarkImprovingRepresentationHydrologic2015;
+@loucksWaterResourceSystems2017], is hindered by manual and
+time-consuming data organization tasks, resulting on time drain,
+disengagement, information overload and human errors
+[@brunsTediousWorkDeveloping2024, @readStateScienceEvolving2021].
 
-The widespread adoption of Geographic Information System (GIS) software has enabled confident decision-making, driven research innovation, and facilitated efficient 
-data management and analysis in various formats [@devantierReviewGisApplications1993; @tsihrintzisUseGeographicInformation1996], particularly in hydrological modeling by integrating spatial and non-spatial data [@martinInterfacingGISWater2005].
-Hydrologic modeling software has evolved by integrating component models and GIS, with developers focusing on improving GUIs and integration with existing tools [@bhattTightlyCoupledGIS2014], 
-whereas specific hydrology-focused software [@rossmanOpenSourcingEPANET2010; @gironasNewApplicationsManual2010] lacks general applicability, highlighting the need for a balanced approach that combines specificity to hydrological
-research questions with general capabilities.
+A directed acyclic graph (DAG) also called directed tree
+[@deoGraphTheoryApplications2016] is one of the best way to represent
+the river network [@rinaldoTreesNetworksHydrology2006;
+@kuhnDesigningLanguageSpatial2015;
+@abed-elmdoustEmergentSpectralProperties2017], characterized by
+directed edges and a hierarchical structure, which offers efficient
+storage and retrieval of large databases compared to relational models
+[@demirOptimizationRiverNetwork2017; @knoxOpensourceDataManager2019].
 
-A directed acyclic graph (DAG) also called directed tree [@deoGraphTheoryApplications2016] is one of the best way to represent the river network [@rinaldoTreesNetworksHydrology2006; @kuhnDesigningLanguageSpatial2015; @abed-elmdoustEmergentSpectralProperties2017], characterized by directed edges and a hierarchical structure, which offers 
[…]
