### 2025-06-19 #8432 @Atreyagaurav
I have removed the `nadi-gis` component, which is in a separate repo, and made this paper only about the domain specific programming language. This should have easier compilation for review as well as shorted paper.

Paper without the metadata is 923 words long.
