### 2025-06-19 c3b5213
add details on tests; paper edits
