+++ paper.md
@@ -15,9 +15,9 @@ authors:
   - name: Patrick A. Ray
     affiliation: 1
 affiliations:
-  - name: Department of Chemical and Environmental Engineering, University of Cincinnati, 601, Engineering Research Center, Cincinnati, OH 45221-0012, United States of America
+  - name: Department of Chemical and Environmental Engineering, University of Cincinnati, OH, USA
     index: 1
-  - name: U. S. Army Engineer Research and Development Center (ERDC), 707 Fourth St., Davis, CA 95616, United States of America
+  - name: U. S. Army Engineer Research and Development Center (ERDC), Davis, CA, USA
     index: 2
 date: 20 May 2025
 bibliography: references.bib
@@ -26,58 +26,59 @@ bibliography: references.bib
 # Summary
 We present Network Analysis and Data Integration (NADI) System, a
 developing software framework designed to facilitate river data
-analysis and visualization. NADI comes with a Domain Specific
-Programming Language (DSL) that has an intuitive syntax for network
-metadata analysis as well as a generalized plugin system to run
-user-defined functions on each node or the whole network. Plugins
-provide seamless integration with other softwares and
-programming languages.
+analysis. NADI comes with a Domain Specific Programming Language (DSL)
+that has an intuitive syntax for network metadata analysis as well as
+a generalized plugin system to run user-defined functions on each node
+or the whole network. Plugins provide seamless integration with other
+softwares and programming languages.
+
+NADI can be used from Command Line Interface (CLI), Graphical User
+Interface (GUI) using the NADI Integrated Development Environment
+(IDE), as a Rust library, or as a Python library allowing users to
+write their own plugins or programs.
 
 # Statement of need
-Hydrologic modeling, which involves integrating diverse data to
-simulate complex processes [@singhMathematicalModelingWatershed2002;
-@clarkImprovingRepresentationHydrologic2015;
-@loucksWaterResourceSystems2017], is hindered by manual and
-time-consuming data organization tasks, resulting on time drain,
-disengagement, information overload and human errors
-[@brunsTediousWorkDeveloping2024, @readStateScienceEvolving2021].
+Hydrological analysis sometimes consists of data that are related to
+points in the river, and frequently with higher correlation or other
+relationship between upstream/downstream points. As many analysis
+starts with building and using that relationship for finding
+inconsistencies in the data, filling missing data, visualization of
+metadata, etc. there is a need for intelligent computational
+assistance to reduce the workload and improve the efficiency and
+reproducibility of research in this field
+[@rosenbergNextFrontierMaking2020]. Specific hydrology-focused
+software [@rossmanOpenSourcingEPANET2010;
+@gironasNewApplicationsManual2010] lack general applicability, while
+general purpose programming languages might not have the succinct and
+intuitive syntax. This highlights the need for a balanced approach
+that combines specificity to hydrological research questions with
+general capabilities.
 
-A directed acyclic graph (DAG) also called directed tree
-[@deoGraphTheoryApplications2016] is one of the best way to represent
-the river network [@rinaldoTreesNetworksHydrology2006;
-@kuhnDesigningLanguageSpatial2015;
-@abed-elmdoustEmergentSpectralProperties2017], characterized by
-directed edges and a hierarchical structure, which offers efficient
-storage and retrieval of large databases compared to relational models
-[@demirOptimizationRiverNetwork2017; @knoxOpensourceDataManager2019].
-
-Domain Specific Programming Language (DSL) has several advantages,
+Domain Specific Programming Language (DSL)s have several advantages
 including improved code readability and maintainability due to the
-DSL's tailored syntax and semantics, increased efficiency through
[…]
