### 2025-06-22 7cfd189
WIP: make series/timeseries with null values
