+++ nadi_core/src/timeseries.rs
@@ -187,6 +187,20 @@ impl TimeSeries {
 #[repr(C)]
 #[derive(StableAbi, Clone, PartialEq, Debug)]
 pub enum Series {
+    Masked(MaskedSeries),
+    Values(SeriesValues),
+}
+
+#[repr(C)]
+#[derive(StableAbi, Clone, PartialEq, Debug)]
+pub struct MaskedSeries {
+    values: SeriesValues,
+    mask: RVec<bool>,
+}
+
+#[repr(C)]
+#[derive(StableAbi, Clone, PartialEq, Debug)]
+pub enum SeriesValues {
     Floats(RVec<f64>),
     Integers(RVec<i64>),
     Strings(RVec<RString>),
@@ -197,7 +211,7 @@ pub enum Series {
     Attributes(RVec<Attribute>),
 }
 
-impl Series {
+impl SeriesValues {
     pub fn floats(v: Vec<f64>) -> Self {
         Self::Floats(v.into())
     }
@@ -281,14 +295,14 @@ impl Series {
 
     pub fn to_attributes(self) -> Vec<Attribute> {
         match self {
-            Series::Floats(v) => v.into_iter().map(Attribute::Float).collect(),
-            Series::Integers(v) => v.into_iter().map(Attribute::Integer).collect(),
-            Series::Strings(v) => v.into_iter().map(Attribute::String).collect(),
-            Series::Booleans(v) => v.into_iter().map(Attribute::Bool).collect(),
-            Series::Dates(v) => v.into_iter().map(Attribute::Date).collect(),
-            Series::Times(v) => v.into_iter().map(Attribute::Time).collect(),
-            Series::DateTimes(v) => v.into_iter().map(Attribute::DateTime).collect(),
-            Series::Attributes(v) => v.into(),
+            Self::Floats(v) => v.into_iter().map(Attribute::Float).collect(),
+            Self::Integers(v) => v.into_iter().map(Attribute::Integer).collect(),
+            Self::Strings(v) => v.into_iter().map(Attribute::String).collect(),
+            Self::Booleans(v) => v.into_iter().map(Attribute::Bool).collect(),
+            Self::Dates(v) => v.into_iter().map(Attribute::Date).collect(),
+            Self::Times(v) => v.into_iter().map(Attribute::Time).collect(),
+            Self::DateTimes(v) => v.into_iter().map(Attribute::DateTime).collect(),
+            Self::Attributes(v) => v.into(),
         }
     }
 
@@ -306,37 +320,37 @@ impl Series {
     }
 }
 
-pub trait FromSeries<'a>: Sized {
-    fn from_series(value: &'a Series) -> Option<&'a [Self]>;
-    fn from_series_mut(value: &'a mut Series) -> Option<&'a mut [Self]>;
-    fn try_from_series(value: &'a Series) -> Result<&'a [Self], String> {
+pub trait FromSeriesValues<'a>: Sized {
+    fn from_series(value: &'a SeriesValues) -> Option<&'a [Self]>;
+    fn from_series_mut(value: &'a mut SeriesValues) -> Option<&'a mut [Self]>;
+    fn try_from_series(value: &'a SeriesValues) -> Result<&'a [Self], String> {
         let ermsg = format!(
             "Incorrect Type: series of `{}` cannot be converted to `{}`",
             value.type_name(),
             type_name::<Self>()
         );
-        FromSeries::from_series(value).ok_or(ermsg)
+        FromSeriesValues::from_series(value).ok_or(ermsg)
     }
-    fn try_from_series_mut(value: &'a mut Series) -> Result<&'a mut [Self], String> {
+    fn try_from_series_mut(value: &'a mut SeriesValues) -> Result<&'a mut [Self], String> {
         let ermsg = format!(
[…]
