### 2025-06-26 f861b0d
WIP: Add docstrings for types/methods the nadi_core
