+++ nadi_core/src/network.rs
@@ -116,10 +116,12 @@ impl HasTimeSeries for Network {
 }
 
 impl Network {
+    /// Iterator for the nodes in the network
     pub fn nodes(&self) -> impl Iterator<Item = &Node> {
         self.nodes.iter().map(|n| &self.nodes_map[n])
     }
 
+    /// Iterator for the edges of the network
     pub fn edges(&self) -> impl Iterator<Item = (&Node, &Node)> + '_ {
         self.edges_ind().map(|(s, e)| {
             (
@@ -129,10 +131,12 @@ impl Network {
         })
     }
 
+    /// Outlet of the network (node with no output)
     pub fn outlet(&self) -> Option<&Node> {
         self.outlet.as_ref().into()
     }
 
+    /// Append the edges from the list, making new nodes if necessary
     pub fn append_edges(&mut self, edges: &[(&str, &str)]) -> Result<(), String> {
         for (start, end) in edges {
             if !self.nodes_map.contains_key(*start) {
@@ -160,17 +164,20 @@ impl Network {
         Ok(())
     }
 
+    /// Create a network with given edges
     pub fn from_edges(edges: &[(&str, &str)]) -> Result<Self, String> {
         let mut network = Self::default();
         network.append_edges(edges)?;
         Ok(network)
     }
 
+    /// Iterator of the edges with nodes' names
     pub fn edges_str(&self) -> impl Iterator<Item = (&str, &str)> + '_ {
         self.edges_ind()
             .map(|(s, e)| (self.nodes[s].as_str(), self.nodes[e].as_str()))
     }
 
+    /// Iterator of the edges with node index
     pub fn edges_ind(&self) -> impl Iterator<Item = (usize, usize)> + '_ {
         self.nodes().filter_map(|n| {
             let n = n.lock();
@@ -181,39 +188,46 @@ impl Network {
         })
     }
 
+    /// Iterator of node names
     pub fn node_names(&self) -> impl Iterator<Item = &str> {
         self.nodes.iter().map(|n| n.as_str())
     }
 
+    /// Nodes iterator in reverse order
     pub fn nodes_rev(&self) -> impl Iterator<Item = &Node> {
         self.nodes.iter().rev().map(|n| &self.nodes_map[n])
     }
 
+    /// Number of nodes in the network
     pub fn nodes_count(&self) -> usize {
         self.nodes.len()
     }
 
+    /// Insert a new node by its name
     pub fn insert_node_by_name(&mut self, name: &str) {
         let node = new_node(self.nodes_count(), name);
         self.nodes_map.insert(name.into(), node);
         self.nodes.push(name.into());
     }
 
+    /// Get a node by index
     pub fn node(&self, ind: usize) -> Option<&Node> {
         self.nodes.get(ind).map(|n| &self.nodes_map[n])
     }
 
+    /// Get a node by name
     pub fn node_by_name(&self, name: &str) -> Option<&Node> {
         self.nodes_map.get(name)
     }
 
+    /// Get a node by name (with error msg on failure)
     pub fn try_node_by_name(&self, name: &str) -> Result<&Node, EvalError> {
         self.nodes_map
             .get(name)
             .ok_or_else(|| EvalError::NodeNotFound(name.to_string()))
     }
 
-    /// use TaskContext::propagation whenever possible
+    /// Get nodes in the given order
     pub fn nodes_order(&self, prop: &PropOrder) -> Vec<Node> {
         match prop {
             PropOrder::Auto | PropOrder::Sequential | PropOrder::OutputFirst => {
@@ -223,7 +237,7 @@ impl Network {
         }
     }
 
-    /// use TaskContext::propagation whenever possible
+    /// Get nodes in the given order and selection
     pub fn nodes_select(
         &self,
         order: &PropOrder,
[…]
