+++ nadi-ide/src/editor.rs
@@ -531,8 +531,8 @@ impl Editor {
 
     pub fn subscription(&self) -> Subscription<Message> {
         Subscription::batch([
-            time::every(Duration::from_secs(1)).map(|_| Message::MaybeSaveEditHist),
-            time::every(Duration::from_secs(2)).map(|_| Message::MaybeParseTasks),
+            time::every(Duration::from_millis(100)).map(|_| Message::MaybeSaveEditHist),
+            time::every(Duration::from_secs(1)).map(|_| Message::MaybeParseTasks),
         ])
     }
 }
