### 2025-07-03 bf3943e
add mechanism to test task script in docstrings, add tests
