+++ nadi_core/src/internal/attrs2.rs
@@ -22,7 +22,12 @@ mod attributes {
     /// from `A` to `D`
     ///
     /// ```task
+    /// network load_str("A -> B\n B -> D");
     /// node[A -> D] set_attrs(a2d = true)
+    ///     ```
+    /// This is equivalent to the following:
+    /// ```task
+    /// node[A->D].a2d = true;
     /// ```
     #[node_func]
     fn set_attrs(
@@ -38,6 +43,11 @@ mod attributes {
     }
 
     /// Retrive attribute
+    ///
+    /// ```task
+    /// network load_str("A -> B\n B -> D");
+    /// node assert_eq(get_attr("NAME"), NAME);
+    /// ```
     #[node_func]
     fn get_attr(
         node: &NodeInner,
@@ -50,6 +60,13 @@ mod attributes {
     }
 
     /// Check if the attribute is present
+    ///
+    /// ```task
+    /// network load_str("A -> B\n B -> D");
+    /// node.x = 90;
+    /// node assert(has_attr("x"))
+    /// node assert(!has_attr("y"))
+    /// ```
     #[node_func]
     fn has_attr(
         node: &NodeInner,
@@ -60,6 +77,17 @@ mod attributes {
     }
 
     /// Return the first Attribute that exists
+    ///
+    /// This is useful when you have a bunch of attributes that might
+    /// be equivalent but are using different names. Normally due to
+    /// them being combined from different datasets.
+    ///
+    /// ```task
+    /// network load_str("A -> B\n B -> D");
+    /// node.x = 90;
+    /// node assert_eq(first_attr(["y", "x"]), 90)
+    /// node assert_eq(first_attr(["x", "NAME"]), 90)
+    /// ```
     #[node_func]
     fn first_attr(
         node: &NodeInner,
@@ -77,6 +105,13 @@ mod attributes {
     }
 
     /// map values from the attribute based on the given table
+    ///
+    /// ```task
+    /// env.val = strmap("Joe", {Dave = 2, Joe = 20});
+    /// env assert_eq(val, 20)
+    /// env.val2 = strmap("Joe", {Dave=2}, default = 12);
+    /// env assert_eq(val2, 12)
+    /// ```
     #[env_func]
     fn strmap(
         /// Value to transform the attribute
@@ -91,6 +126,28 @@ mod attributes {
     }
 
     /// if else condition with multiple attributes
+    ///
+    /// ```task
+    /// network load_str("a -> b");
+    /// env.some_condition = true;
+    /// node set_attrs_ifelse(
+    ///     env.some_condition,
+    ///     val1 = [1, 2],
+    ///     val2 = ["a", "b"]
+    /// );
+    /// env assert_eq(nodes.val1, [1, 1])
+    /// env assert_eq(nodes.val2, ["a", "a"])
+    /// ```
+    /// This is equivalent to using the if-else expression directly,
+    ///
+    /// ```task
+    /// node.val1 = if (env.some_condition) {1} else {2};
+    /// env assert_eq(nodes.val1, [1, 1])
+    /// ```
+    ///
+    /// Furthermore if-else expression will give a lot more
+    /// flexibility than this function in normal use cases. But this
+    /// function is useful when you have to do something in a batch.
     #[node_func]
     fn set_attrs_ifelse(
         node: &mut NodeInner,
@@ -110,6 +167,15 @@ mod attributes {
     }
 
     /// Set node attributes based on string templates
+    ///
+    /// This renders the template for each node, then it sets the
+    /// values from the rendered results.
+    ///
+    /// ```task
+    /// network load_str("a -> b");
+    /// node set_attrs_render(val1 = "Node: {_NAME}");
[…]
