### 2025-07-05 ee8be2b
some docstrings, attribute viewer shows array/table in nested fmt
