+++ nadi-ide/src/attributes.rs
@@ -1,57 +1,217 @@
+use crate::icons;
 use iced::widget::{Column, column, container, horizontal_space, row, scrollable, text};
 use iced::{Element, Fill, Length, color};
-use nadi_core::attrs::AttrMap;
+use nadi_core::attrs::{AttrMap, Attribute};
 
+/// State of Attribute Viewer
 #[derive(Default)]
 pub struct AttrView {
     label: String,
     // attributes: AttrMap,
-    values: Vec<(String, String, String)>,
+    values: Vec<Attr>,
+}
+
+#[derive(Clone, Debug)]
+pub enum Message {
+    Toggle(String),
+    CollapseAll,
+    ExpandAll,
+}
+
+/// Each attribute in the attribute viewer
+struct Attr {
+    id: String,
+    name: String,
+    ty: String,
+    val: AttrVal,
+    expanded: bool,
+}
+
+/// Different type of values for Attribute
+enum AttrVal {
+    Single(String),
+    Multiple(Vec<Attr>),
+}
+
+impl Attr {
+    fn new(attr: &Attribute, name: &str, pre: &str) -> Attr {
+        let id = if pre.is_empty() {
+            name.to_string()
+        } else {
+            format!("{pre}.{name}")
+        };
+        let (expanded, ty, val) = match attr {
+            Attribute::Array(ar) => (
+                ar.len() < 5,
+                format!("Array [{}]", ar.len()),
+                AttrVal::Multiple(
+                    ar.iter()
+                        .enumerate()
+                        .map(|(i, a)| Attr::new(a, &i.to_string(), &id))
+                        .collect(),
+                ),
+            ),
+            Attribute::Table(am) => (
+                am.len() < 5,
+                format!("Table [{}]", am.len()),
+                AttrVal::Multiple(am.iter().map(|a| Attr::new(a.1, a.0, &id)).collect()),
+            ),
+            a => (
+                true,
+                a.type_name().to_string(),
+                AttrVal::Single(a.to_string()),
+            ),
+        };
+        Attr {
+            id: id.clone(),
+            name: name.to_string(),
+            ty,
+            val,
+            expanded,
+        }
+    }
+
+    fn toggle(&mut self) {
+        self.expanded = !self.expanded;
+    }
+
+    fn set_expanded(&mut self, exp: bool) {
+        self.expanded = exp;
+        let mem = match self.val {
+            AttrVal::Single(_) => return,
+            AttrVal::Multiple(ref mut ar) => ar,
+        };
+
+        for i in mem {
+            i.set_expanded(exp);
+        }
+    }
+
+    fn toggle_internal(&mut self, id: &str) {
+        let mem = match self.val {
+            AttrVal::Single(_) => return,
+            AttrVal::Multiple(ref mut ar) => ar,
+        };
+
+        for i in mem {
+            if id.starts_with(&i.id) {
+                if id.len() > i.id.len() {
+                    i.toggle_internal(&id);
+                } else {
+                    i.toggle();
+                }
+            }
+        }
+    }
+
+    fn view(&self) -> Element<'_, Message> {
+        let header = container(
+            row![
+                text(&self.name).width(Length::FillPortion(3)),
+                text(&self.ty)
+                    .width(Length::FillPortion(2))
+                    .align_x(iced::alignment::Horizontal::Right),
+                if self.expanded {
+                    icons::action(
+                        icons::resize_small_icon(),
+                        "Collapse",
+                        Some(Message::Toggle(self.id.clone())),
+                    )
+                } else {
+                    icons::action(
+                        icons::resize_full_icon(),
+                        "Expand",
+                        Some(Message::Toggle(self.id.clone())),
+                    )
+                }
+            ]
[…]
