### 2025-07-05 a09b14d
add functions to subset network by largest connected nodes
Helps us filter extra fragments
