### 2025-07-15 6002c39
add video instruction links; add syntax sugar for var default value
