+++ nadi_core/src/parser/expressions.rs
@@ -13,7 +13,7 @@ use nom::{
 
 pub fn expression<'a, 'b>(inp: &'a [Token<'b>]) -> MatchRes<'a, 'b, Expression> {
     alt((
-        map(input_variable, Expression::Variable),
+        input_variable,
         map(attribute, Expression::Literal),
         map(function_call, Expression::Function),
         uni_operator_expr,
@@ -140,16 +140,34 @@ pub fn variable_type<'a, 'b>(inp: &'a [Token<'b>]) -> MatchRes<'a, 'b, VarType> 
     }
 }
 
-pub fn input_variable<'a, 'b>(inp: &'a [Token<'b>]) -> MatchRes<'a, 'b, InputVar> {
+pub fn input_variable<'a, 'b>(inp: &'a [Token<'b>]) -> MatchRes<'a, 'b, Expression> {
     map(
         tuple((
             opt(terminated(variable_type, dot)),
             dot_variable,
-            opt(question),
+            opt(maybe_space(pair(
+                question,
+                opt(maybe_space(alt((expression, expression_group)))),
+            ))),
         )),
         |(vt, mut v, q)| {
             let name = v.pop().expect("There should be at least one var");
-            InputVar::new(vt, v, name, q.is_some())
+            if let Some((_, val)) = q {
+                if let Some(val) = val {
+                    let cond = Expression::Variable(InputVar::new(
+                        vt.clone(),
+                        v.clone(),
+                        name.clone(),
+                        true,
+                    ));
+                    let var = Expression::Variable(InputVar::new(vt, v, name, false));
+                    Expression::IfElse(Box::new(cond), Box::new(var), Box::new(val))
+                } else {
+                    Expression::Variable(InputVar::new(vt, v, name, true))
+                }
+            } else {
+                Expression::Variable(InputVar::new(vt, v, name, false))
+            }
         },
     )(inp)
 }
