### 2025-07-17 #8432 @kthyng
@Atreyagaurav
* Is this documentation specific to (or at least thoroughly covers) NADI? https://nadi-system.github.io/
* I'm struggling to understand what your package is about. Can you add more high level language to the repo so people who aren't already experts understand the motivation? For example, your description is "Mono Repo for nadi_core, cli, gui" but that isn't clear.
