+++ README.md
@@ -1,10 +1,25 @@
 # Nadi System
 
-Collection of Utilities to do Network Analysis and Data Integration. 
+Collection of Utilities to do Network Analysis and Data Integration. This is made targeting analysis of point data in a river system but it should work for any network analysis that work on directed tree network.
 
-For more details, refer to the [NADI Book](https://nadi-system.github.io/)
+The main component of the NADI System is the DSL. It can be used for network based data analysis. The repo contains the tools for using nadi through Command Line Interface (CLI), Integrated Development Environment (IDE), `mdbook` (documentation writing tool), or as a python library.
+
+The overall NADI workflow is shown below. This repository contains the components for the Network Analysis part of the system.
+
+![Nadi Workflow](nadi-workflow-simple.svg)
+
+## Theory
+The data associated with points in a river system is loaded as a network, and along side the functions loaded through the plugins, the Domain Specific Programming language (DSL) can be used to run different arithmetic, logical, or functional analysis on the metadata of the points.
+
+This allows us to do network based data analysis quickly, with intuitive syntax compared to using general purpose programming language like Python. But the general purpose languages have more flexibility, so a way to couple with them is provided through the plugin mechanism.
+
+## Further Reading
+Please refer to the [NADI Book](https://nadi-system.github.io/) for details on the key concepts, the syntax of the language, and other components of the NADI System.
+
+For developers the API documentation is on [docs.rs](https://docs.rs/nadi_core/latest/nadi_core/index.html).
 
 # Installation
+
 Prebuilt binaries are available for windows in the releases page. Use the `nadi-ide` binary for the GUI, refer to Nadi Book for details on other binaries.
 
 ![Nadi IDE Screenshot](ide-screenshot.png)
