### 2025-07-17 #8432 @Atreyagaurav
@kthyng Thank you for the questions.

>     * Is this documentation specific to (or at least thoroughly covers) NADI? https://nadi-system.github.io/

Yes, that is a documentation specific to NADI. It is written as a user guide. It should cover the installation, motivation, concepts and examples for the system, then the reference to the functions from plugins and other details.

Documentation for individual functions are available in their respective pages: https://nadi-system.github.io/plugins/command.html

The developer reference is automatically generated through doc strings and available at docs.rs website, that's where the rust package documentation is generally available for rust developers. I added a reference to this on the repository.

https://docs.rs/nadi_core/latest/nadi_core/index.html

Here is an example page explaining the `Attribute` enum: https://docs.rs/nadi_core/latest/nadi_core/attrs/enum.Attribute.html

>     * I'm struggling to understand what your package is about. Can you add more high level language to the repo so people who aren't already experts understand the motivation? For example, your description is "Mono Repo for nadi_core, cli, gui" but that isn't clear.

Thank you. More details are available on the Nadi Book, but I added a few details on the repo description as well as the readme. Please let me know if I need to improve it further.
