### 2025-07-18 #8432 @mengqi-z
👋 @Atreyagaurav @kvenkman @jhwohlgemuth - Would any of you be interested and available to review a submission for JOSS? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
