### 2025-07-19 #8432 @jhwohlgemuth
> 👋 [@Atreyagaurav](https://github.com/Atreyagaurav) [@kvenkman](https://github.com/kvenkman) [@jhwohlgemuth](https://github.com/jhwohlgemuth) - Would any of you be interested and available to review a submission for JOSS? JOSS uses a checklist-driven review process, which you can find here: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

I can do it! 🫡
