### 2025-07-20 82b1794
add short help of functions on signature; combine error and status
