### 2025-07-20 16bb59e
add nadi-serve with basic UI to run independent Nadi Tasks
