### 2025-07-21 0611555
feat: [mdbook] add the option to continue the task execution
