+++ CONTRIBUTING.md
@@ -1,0 +1,31 @@
+Thank you for considering contributing to the NADI System. NADI System consists of multiple components so there are variety of ways to contribute.
+
+Please note that you do not need experience with Rust for you to contribute to the NADI System. There are few categories of contributions outlined below.
+
+- User: you only need to use the system,
+- Dev: you should be able to write programs in Rust,
+- Doc: you can help with documentation (coding in NADI might be preferred, not in rust),
+- Extra: Some other skills related to tech of your choice
+
+# [User] Sharing NADI
+You can contribute to NADI by sharing it with researchers and other users that might be interested or benefit from the unique paradigm it offers.
+
+# [User] Bug Reporting
+As NADI is currently in rapid development stage, there is bound to be a few bugs. If you have encountered a bug while using NADI system, please make an issue, with details on the version, how you installed NADI, and the Operating System.
+
+# [Doc] Documenting NADI
+You can also help with testing the NADI codes from Nadi book, as well as adding more examples, or other use cases you can find. Please make an issue or pull request about this in the [Nadi Book repository](https://github.com/Nadi-System/nadi-book). Please feel free to make issues regarding any troubles you had while reading the books. As well as any infomation you think is important to mention in the Book.
+
+Besides the book, you can also document it in your own platforms, make posts or videos demostrating your use of NADI and other ways people might discover NADI and its features.
+
+# [User] Feature Request
+If you have a feature you think would be a good match with nadi please make an issue as well, please start it with [Feature]. To make sure there is no wasted effort, please wait for the confirmation from the maintainer before you start implementing the feature. Although as it is open source you are free to add your own spin in a fork and use that.
+
+# [Dev] Plugin Development
+This might be the easy way to get into development of NADI. You can write plugins following the instructions from Nadi Book. The `nadi` command will generate a template plugin for you, and you can change things there, compile it, and test it from the NADI system. Being able to develop plugins means you can expand the use case of NADI based on your demands.
+
+# [Extra] Interoperability with other Program
+NADI is still incomplete, and to not reinvent the wheel, NADI provides plugins that can interface with other programs. If you use some programs that you think would benefit from interface with NADI, write a plugin for it. And if it works well you can share it with the community. We recommend NADI Plugins to be open source so that users/community can inspect the code before deciding to use it.
+
+# [Dev] Nadi Core development
+If your expertise lies in development in rust, programming language development, IDE features, and such please feel free to open an issue with what you want to work on. Refer to the future works page in the book for some ideas, or come up with your own ideas. You can also start with requesting/discussing a feature, and then develop it later on.
