### 2025-07-24 #8432 @dorchard
Whilst I'm knowledgable in DSLs and their associated theory and practice, I don't know anything about the domain and am time limited over the summer, so maybe there is someone better with more time at the moment. Sorry.
