### 2025-07-24 24659f6
Deactivate the tasks syntax checker for now due to performace issues
