### 2025-07-25 9d2dc88
add line/column information to eval error through data types
