+++ nadi_core/src/parser/tokenizer.rs
@@ -1,42 +1,95 @@
+use crate::expressions::TaskPosition;
 use crate::parser::string::parse_string;
 use crate::parser::{ParseError as TaskParseError, ParseErrorType};
 use crate::tasks::TaskKeyword;
 use colored::Colorize;
 use nadi_core::attrs::{Attribute, Date, DateTime, Time};
 use nom::{
+    IResult,
     branch::alt,
     bytes::complete::{is_not, tag},
     character::complete::{alpha1, alphanumeric1, anychar, char, digit1, one_of},
     combinator::{map, opt, recognize},
-    error::{context, VerboseError},
+    error::{VerboseError, context},
     multi::{many0, many1},
     sequence::{pair, preceded, terminated, tuple},
-    IResult,
 };
 use std::str::FromStr;
 
 #[derive(Clone, PartialEq, Debug)]
 pub struct Token<'a> {
+    /// Token Type
     pub ty: TaskToken,
+    /// Contents of the token
     pub content: &'a str,
+    /// start position of token (line, col)
+    pub start: (usize, usize),
 }
 
-impl<'a> Token<'a> {
-    fn new(ty: TaskToken, content: &'a str) -> Self {
-        Self { ty, content }
+impl<'a> TaskPosition for Token<'a> {
+    fn position(&self) -> (usize, usize) {
+        self.start
     }
 }
 
-pub fn check_tokens(tokens: &[Token]) -> Result<(), TaskParseError> {
-    let mut data = tokens.split(|t| !t.ty.is_valid());
-    let valid = data.next().unwrap();
-    match data.next() {
-        Some(_) => Err(TaskParseError::new(
-            tokens,
-            &tokens[valid.len()..],
-            ParseErrorType::InvalidToken,
-        )),
-        None => Ok(()),
+impl<'a, 'b> TaskPosition for &'b [Token<'a>] {
+    fn position(&self) -> (usize, usize) {
+        if let Some(t) = self.get(0) {
+            t.position()
+        } else {
+            (0, 0)
+        }
+    }
+}
+
+impl<'a> Token<'a> {
+    fn new(raw: RawToken<'a>, start: (usize, usize)) -> Self {
+        Self {
+            ty: raw.ty,
+            content: raw.content,
+            start,
+        }
+    }
+    pub fn validate(tokens: Vec<RawToken<'a>>) -> Result<Vec<Self>, TaskParseError> {
+        let mut data = tokens.split(|t| !t.ty.is_valid());
+        let valid = data.next().unwrap();
+        if data.next().is_some() {
+            return Err(TaskParseError::raw(
+                &tokens,
+                &tokens[valid.len()..],
+                ParseErrorType::InvalidToken,
+            ));
+        }
+        let mut line = 1;
+        let mut col = 1;
+        let tokens = tokens
+            .into_iter()
+            .map(|t| {
+                let start = (line, col);
+                if t.ty == TaskToken::NewLine {
+                    line += 1;
+                    col = 1;
+                } else {
+                    col += t.content.len();
+                }
+                Token::new(t, start)
+            })
+            .collect();
+        Ok(tokens)
+    }
+}
+
+#[derive(Clone, PartialEq, Debug)]
+pub struct RawToken<'a> {
+    /// Token Type
+    pub ty: TaskToken,
+    /// Contents of the token
+    pub content: &'a str,
+}
+
+impl<'a> RawToken<'a> {
+    fn new(ty: TaskToken, content: &'a str) -> Self {
+        Self { ty, content }
     }
 }
 
[…]
