### 2025-07-25 fa356e1
new-version in cargo + check for nadi keywords in function argument
The parser simply won't work if you try to use keyword as variables so
we need to check it during compile time and stop plugin developers
from using them
