+++ nadi-serve/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
 name = "nadi-serve"
-version = "0.1.0"
+version = "0.8.0"
 edition = "2024"
 
 [dependencies]
 anyhow = "1.0.89"
 clap = { version = "4.5.17", features = ["derive"] }
-nadi_core = {version = "0.7.1", path="../nadi_core", features=["functions", "parser"]}
+nadi_core = {version = "0.8.0", path="../nadi_core", features=["functions", "parser"]}
 rocket = { version = "0.5.1", features = ["json"] }
 serde = "1.0.219"
