+++ nadi_core/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nadi_core"
-version = "0.7.1"
+version = "0.8.0"
 edition = "2021"
 description = "Core library for Nadi systems, for use by plugins"
 repository = "https://github.com/Nadi-System/nadi-system"
@@ -17,7 +17,7 @@ anyhow = "1.0.86"
 chrono = {version = "0.4.38", optional = true}
 nom = {version = "7.1.3", optional = true}
 string-template-plus = "0.4.3"
-nadi_plugin = {version = "0.7.1", path="../nadi_plugin"}
+nadi_plugin = {version = "0.8.0", path="../nadi_plugin"}
 colored = "2.1.0"
 subprocess = "0.2.9"
 number_range = "0.3.2"
