+++ nadi_plugin/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nadi_plugin"
-version = "0.7.1"
+version = "0.8.0"
 edition = "2021"
 description = "Macro library to write plugins for nadi system using nadi_core"
 repository = "https://github.com/Nadi-System/nadi_plugin"
