### 2025-07-25 7b8c637
add channels for communications; remove duplicate plugins loading
