### 2025-07-28 #8432 @mengqi-z
Hi @Atreyagaurav,

Thanks for checking and suggesting reviewers. You're absolutely right. It's quite rare to find reviewers with expertise in both the domain and the language. I'm planning to reach out to a few more potential reviewers who have a relevant background, and hopefully one of them will be available and willing to review.

If we’re unable to find a suitable reviewer within the JOSS reviewer pool, it’s also common practice to look beyond it through professional networks or by identifying potential experts from relevant publications.
