### 2025-07-28 #8432 @mengqi-z
👋 @JosiahParry @narnolddd @lolpro11 @amcandio - Would any of you be willing and available to review this submission for JOSS? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
