### 2025-07-28 #8655 @editorialbot
[REVIEW]: NADI -- Network Analysis and Data Integration with a Domain Specific Programming Language
**Submitting author:** @atreyagaurav (<a href="http://orcid.org/0000-0002-0234-2165">Gaurav Atreya</a>)
**Repository:** https://github.com/Nadi-System/nadi-system
**Branch with paper.md** (empty if default branch):
**Version:** 0.7.2
