### 2025-07-28 #8655 @Atreyagaurav
Thank you very much @mengqi-z for starting the review.

Thank you, @jhwohlgemuth and @lolpro11 for agreeing to review this paper. :)

You'd probably want the details on the installation/usage instructions clearly available, so please let me know if you have any trouble with it, I'll edit the documentation accordingly. Some parts of the documentation are work in progress, for the core documentation and examples should work.
