### 2025-07-28 a2ff9db
change DAG into Rooted tree in the paper, and grammar
