+++ paper.md
@@ -24,21 +24,21 @@ bibliography: references.bib
 ---
 
 # Summary
-We present Network Analysis and Data Integration (NADI) System, a
+We present the Network Analysis and Data Integration (NADI) System, a
 developing software framework designed to facilitate river data
-analysis. NADI comes with a Domain Specific Programming Language (DSL)
-that has an intuitive syntax for network metadata analysis as well as
-a generalized plugin system to run user-defined functions on each node
-or the whole network. Plugins provide seamless integration with other
-softwares and programming languages.
+analysis. NADI System comes with a Domain Specific Programming
+Language (DSL) that has an intuitive syntax for network metadata
+analysis as well as a plugin system to run user-defined functions on
+each node or the whole network. Plugins provide seamless integration
+with other softwares and programming languages.
 
-NADI can be used from Command Line Interface (CLI), Graphical User
+NADI can be used from the Command Line Interface (CLI), Graphical User
 Interface (GUI) using the NADI Integrated Development Environment
-(IDE), as a Rust library, or as a Python library allowing users to
+(IDE), as a Rust library, or as a Python library, allowing users to
 write their own plugins or programs.
 
 # Statement of need
-Hydrological analysis sometimes consists of data that are related to
+Hydrological analysis sometimes consists of data that is related to
 points in the river, and frequently with higher correlation or other
 relationship between upstream/downstream points. As many analysis
 starts with building and using that relationship for finding
@@ -47,40 +47,43 @@ metadata, etc. there is a need for intelligent computational
 assistance to reduce the workload and improve the efficiency and
 reproducibility of research in this field
 [@rosenbergNextFrontierMaking2020]. Specific hydrology-focused
-software [@rossmanOpenSourcingEPANET2010;
+softwares [@rossmanOpenSourcingEPANET2010;
 @gironasNewApplicationsManual2010] lack general applicability, while
 general purpose programming languages might not have the succinct and
 intuitive syntax. This highlights the need for a balanced approach
 that combines specificity to hydrological research questions with
 general capabilities.
 
-Domain Specific Programming Language (DSL)s have several advantages
+Domain Specific Programming Languages (DSLs) have several advantages
 including improved code readability and maintainability due to the
-DSL's tailored syntax and semantics as well as increased efficiency
-through optimized algorithms and data structures
+tailored syntax and semantics, as well as increased efficiency through
+optimized algorithms and data structures
 [@mernikWhenHowDevelop2005]. Although Graphviz has made great progress
 on the visualization of graph [@gansnerOpenGraphVisualization2000;
 @ellsonGraphvizDynagraphStatic2004], it does not have the analytical
 capabilities. Hydrolang has the capabilities of analysis and
-visualization for hydrological applications, but it is web based
+visualization for hydrological applications, but it is web-based
 [@erazoramirezHydroLangOpensourceWebbased2022]. Languages made for
-spatial analysis are mostly working with grid based system
-[@pullarMapScriptMapAlgebra2001; @kuhnDesigningLanguageSpatial2015].
+grid-based spatial analysis [@pullarMapScriptMapAlgebra2001;
+@kuhnDesigningLanguageSpatial2015] are not suitable for values like
+streamflow which are partially spatially continuous along the river,
+but not in 2D/3D space.
 
-We present NADI that can load a river network as a Directed Acyclic
-Graph [@deoGraphTheoryApplications2016] which is known to be one of
-the best way to represent the river network
+We present NADI System that can load a river network as a Rooted Tree
+Graph [@deoGraphTheoryApplications2016] --- which is known to be one
+of the best ways to represent the river network
 [@rinaldoTreesNetworksHydrology2006;
 @kuhnDesigningLanguageSpatial2015;
[…]
