### 2025-07-28 #8655 @Atreyagaurav
@lolpro11 Thank you for the quick review.

Yeah, it is general in a sense that you should be able to use it for anything that has similar network structure and dependencies in the data. I wanted to make it work to as many cases as possible with that constraint. But my domain, and the domain it is extensively made for and tested in is related tor River network, so the emphasis is on that.

Do you mean I should highlight the general applicability (different domains) in the documentation?

I am looking for some specific examples in other domains that I could add as an example similar to this: https://nadi-system.github.io/example/ohio-validation.html
