### 2025-07-28 #8655 @Atreyagaurav
Thank you!

How to extend the system is in the Plugin Developer Guide: https://nadi-system.github.io/plugin-dev/external-plugins.html
And some example plugins are here: https://github.com/Nadi-System/nadi-plugins-rust

But it is just technical steps for someone who already knows what they want, so I'll look more into the isochrone analysis, and other examples, and add them as examples people can implement. Giving higher level details on what could be done using this would definitely be a good addition.
