### 2025-07-29 4e780b6
function alias only for internal functions; add hook tasks
