+++ nadi_core/src/parser/components.rs
@@ -108,6 +108,7 @@ one_token!(kw_else, TaskToken::Keyword(TaskKeyword::Else));
 one_token!(kw_while, TaskToken::Keyword(TaskKeyword::While));
 one_token!(kw_in, TaskToken::Keyword(TaskKeyword::In));
 one_token!(kw_match, TaskToken::Keyword(TaskKeyword::Match));
+one_token!(kw_hook, TaskToken::Keyword(TaskKeyword::Hook));
 one_token!(kw_help, TaskToken::Keyword(TaskKeyword::Help));
 one_token!(kw_end, TaskToken::Keyword(TaskKeyword::End));
 one_token!(kw_exit, TaskToken::Keyword(TaskKeyword::Exit));
