### 2025-07-29 #8655 @lolpro11
Well that looks good then, maybe link the docs in the final paper as a citation or as an inline ref?
