### 2025-07-30 da04c31
Grammar + add nadi book reference on the paper
