+++ references.bib
@@ -16,8 +16,7 @@
   abstract = {Characterization of river drainage networks has been a subject of research for many years. However, most previous studies have been limited to quantities which are loosely connected to the topological properties of these networks. In this work, through a graph-theoretic formulation of drainage river networks, we investigate the eigenvalue spectra of their adjacency matrix. First, we introduce a graph theory model for river networks and explore the properties of the network through its adjacency matrix. Next, we show that the eigenvalue spectra of such complex networks follow distinct patterns and exhibit striking features including a spectral gap in which no eigenvalue exists as well as a finite number of zero eigenvalues. We show that such spectral features are closely related to the branching topology of the associated river networks. In this regard, we find an empirical relation for the spectral gap and nullity in terms of the energy dissipation exponent of the drainage networks. In addition, the eigenvalue distribution is found to follow a finite-width probability density function with certain skewness which is related to the drainage pattern. Our results are based on optimal channel network simulations and validated through examples obtained from physical experiments on landscape evolution. These results suggest the potential of the spectral graph techniques in characterizing and modeling river networks.},
   copyright = {2017 The Author(s)},
   langid = {english},
-  keywords = {Geomorphology,Hydrology},
-  file = {/home/gaurav/Zotero/storage/BT2T5SCH/Abed-Elmdoust et al. - 2017 - Emergent spectral properties of river network topology an optimal channel network approach.pdf}
+  keywords = {Geomorphology,Hydrology}
 }
 
 @article{alabyanTypesRiverChannel1998,
@@ -35,8 +34,7 @@
   abstract = {River channel patterns are thought to form a morphological continuum. This continuum is two-dimensional, defined by plan features of which there are three (straight, meandering, branching), and structural levels of fluvial relief of which there are also three (floodplain, flood channel, low-water channel). Combinations of these three categories define the diversity of patterns. One of the most important factors in channel development is stream power, defined by water discharge and river slope. The greater the stream power, the stronger the branching tendency, but threshold values of stream power are different for the three different hierarchical levels of channel relief. The critical stream power values and hydrological regime together define the channel pattern, and analysis of the pattern type can be undertaken using effective discharge curves. {\copyright} 1998 John Wiley \& Sons, Ltd.},
   copyright = {Copyright {\copyright} 1998 John Wiley \& Sons, Ltd.},
   langid = {english},
-  keywords = {anastomosis,braiding,channel pattern,continuum concept,effective discharge curve,meandering},
-  file = {/home/gaurav/Zotero/storage/8KEMMMI3/Alabyan and Chalov - 1998 - Types of river channel patterns and their natural controls.pdf;/home/gaurav/Zotero/storage/JD9N9E5Z/(SICI)1096-9837(199805)235467AID-ESP8613.0.html}
+  keywords = {anastomosis,braiding,channel pattern,continuum concept,effective discharge curve,meandering}
 }
 
 @article{arulGraphTheoryAlgorithms2023,
@@ -47,8 +45,7 @@
   journal = {E3S Web of Conferences},
   volume = {399},
   doi = {10.1051/e3sconf/202339908002},
[…]
