### 2025-07-30 4067ff5
Test: remove nadi-py from Cargo.toml
