### 2025-07-30 f86b7c1
Revert "Test: remove nadi-py from Cargo.toml": maturin fails
This reverts commit 4067ff5accb725f7efde53090d2579c595602181.
