### 2025-07-30 dda48f0
add path to nadi-py
