+++ .github/workflows/nadi-py.yml
@@ -10,6 +10,7 @@ on:
     branches:
       - main
       - master
+      - test-ci
     tags:
       - '*'
   pull_request:
@@ -44,6 +45,7 @@ jobs:
           python-version: 3.x
       - name: Build wheels
         uses: PyO3/maturin-action@v1
+	working-directory: nadi-py
         with:
           target: ${{ matrix.platform.target }}
           args: --release --out dist --find-interpreter --manifest-path nadi-py/Cargo.toml
@@ -51,6 +53,7 @@ jobs:
           manylinux: auto
       - name: Upload wheels
         uses: actions/upload-artifact@v4
+	working-directory: nadi-py
         with:
           name: wheels-linux-${{ matrix.platform.target }}
           path: dist
@@ -75,6 +78,7 @@ jobs:
           python-version: 3.x
       - name: Build wheels
         uses: PyO3/maturin-action@v1
+	working-directory: nadi-py
         with:
           target: ${{ matrix.platform.target }}
           args: --release --out dist --find-interpreter --manifest-path nadi-py/Cargo.toml
@@ -82,6 +86,7 @@ jobs:
           manylinux: musllinux_1_2
       - name: Upload wheels
         uses: actions/upload-artifact@v4
+	working-directory: nadi-py
         with:
           name: wheels-musllinux-${{ matrix.platform.target }}
           path: dist
@@ -103,12 +108,14 @@ jobs:
           architecture: ${{ matrix.platform.target }}
       - name: Build wheels
         uses: PyO3/maturin-action@v1
+	working-directory: nadi-py
         with:
           target: ${{ matrix.platform.target }}
           args: --release --out dist --find-interpreter --manifest-path nadi-py/Cargo.toml
           sccache: ${{ !startsWith(github.ref, 'refs/tags/') }}
       - name: Upload wheels
         uses: actions/upload-artifact@v4
+	working-directory: nadi-py
         with:
           name: wheels-windows-${{ matrix.platform.target }}
           path: dist
@@ -129,12 +136,14 @@ jobs:
           python-version: 3.x
       - name: Build wheels
         uses: PyO3/maturin-action@v1
+	working-directory: nadi-py
         with:
           target: ${{ matrix.platform.target }}
           args: --release --out dist --find-interpreter --manifest-path nadi-py/Cargo.toml
           sccache: ${{ !startsWith(github.ref, 'refs/tags/') }}
       - name: Upload wheels
         uses: actions/upload-artifact@v4
+	working-directory: nadi-py
         with:
           name: wheels-macos-${{ matrix.platform.target }}
           path: dist
@@ -145,11 +154,13 @@ jobs:
       - uses: actions/checkout@v4
       - name: Build sdist
         uses: PyO3/maturin-action@v1
+	working-directory: nadi-py
         with:
           command: sdist
           args: --out dist --manifest-path nadi-py/Cargo.toml
       - name: Upload sdist
         uses: actions/upload-artifact@v4
+	working-directory: nadi-py
         with:
           name: wheels-sdist
           path: dist
