### 2025-07-31 eabb7e8
Change date on nadi book reference
