+++ references.bib
@@ -562,7 +562,7 @@
   shorttitle = {Nadi {{Book}}},
   author = {Atreya, Gaurav},
   year = {2025},
-  month = aug,
+  month = jul,
   series = {Nadi {{Book}}},
   edition = {0.7.0},
   url = {https://nadi-system.github.io/},
