### 2025-07-31 #8655 @Atreyagaurav
@lolpro11 I have added the User Manual as a reference, the extensibility section has it cited as `Atreya (2025)`.

> Atreya, G. (2025). Network Analysis and Data Integration (NADI) System: User Manual (0.7.0 ed.). https://nadi-system.github.io/

I have also updated the User Manual, and improved the PDF version: https://nadi-system.github.io/data/nadi-book.pdf
