### 2025-08-01 Nadi-System/nadi-system#1 @amcandio
JOSS review: dead links in Nadi book
Hi! I'm starting to read the book and I've seen some dead links. For example in https://nadi-system.github.io/how-to.html:

```
If you are in a hurry, but this is your first time reading this book, at least read the [Core Concepts](https://nadi-system.github.io/intro.html), then refer to the section you are interested in.
```

Can you validate all links in the doc are ok?

https://github.com/openjournals/joss-reviews/issues/8655
