### 2025-08-01 Nadi-System/nadi-system#2 @amcandio
JOSS review: software citation missing
Please make sure that all software dependencies are correctly cited in the paper. For example, Python and Rust should be cited.

https://github.com/openjournals/joss-reviews/issues/8655
