### 2025-08-01 Nadi-System/nadi-system#3 @amcandio
JOSS review: inconsistent capitalization in NADI
Please make sure that all mentions to NADI in paper and docs are consistent. We can see "Nadi", "NADI" and "NADI system".

https://github.com/openjournals/joss-reviews/issues/8655
