### 2025-08-01 Nadi-System/nadi-system#4 @amcandio
JOSS review: replace or clarify "intuitive" in paper
The paper uses the word **"intuitive"** to describe the DSL syntax. This is a subjective term and not ideal for academic writing.

Maybe you can explain *why* it is intuitive (e.g., simple expressions for upstream sums) or replace it with clearer wording like:
  - “domain-specific syntax”
  - “syntax familiar to hydrologists”
  - “succinct and readable syntax”

This makes the text clearer and more objective.

https://github.com/openjournals/joss-reviews/issues/8655
