### 2025-08-01 Nadi-System/nadi-system#6 @amcandio
JOSS review: fair evaluation of DSL choice
The paper introduces a custom Domain Specific Language (DSL) but doesn’t evaluate the tradeoffs thoroughly. Paper should be clear about the trade-offs: steep learning curves, tooling overhead, adoption.

https://github.com/openjournals/joss-reviews/issues/8655
