### 2025-08-01 Nadi-System/nadi-system#7 @amcandio
JOSS review: writing improvements
The manuscript would benefit from careful editing to improve clarity, consistency, and tone. Some sentences are awkwardly phrased or grammatically incorrect—for example, “Graphviz has made great progress on the visualization of graph” should be revised for clarity and grammar. The tone is occasionally informal (e.g., “comes with,” “you can,” “easier”), which may be better suited to documentation than an academic paper.

https://github.com/openjournals/joss-reviews/issues/8655
