### 2025-08-01 Nadi-System/nadi-system#8 @amcandio
JOSS review: writing improvements on language description
The description of the language constructs would benefit from greater clarity and structure. Terms such as "task context," "propagation," and "mutable calls" are introduced without clear definitions or sufficient explanation.

https://github.com/openjournals/joss-reviews/issues/8655
