### 2025-08-01 e7ce84a
WIP: fixing problems while loading multiple network in append mode
