### 2025-08-01 0abbf90
add citation for Rust; replace 'intuitive'
Related: https://github.com/Nadi-System/nadi-system/issues/2
Related: https://github.com/Nadi-System/nadi-system/issues/4
