+++ references.bib
@@ -438,6 +438,19 @@
   langid = {english}
 }
 
+@book{klabnikRustProgrammingLanguage2023,
+  title = {The {{Rust}} Programming Language},
+  author = {Klabnik, Steve and Nichols, Carol},
+  year = {2023},
+  edition = {2nd edition},
+  publisher = {No Starch Press},
+  address = {San Francisco},
+  abstract = {"The Rust Programming Language, 2nd Edition is the official guide to Rust 2021: an open source systems programming language that will help you write faster, more reliable software. Rust provides control of low-level details along with high-level ergonomics, allowing you to improve productivity and eliminate the hassle traditionally associated with low-level languages. Klabnik and Nichols, alumni of the Rust Core Team, share their knowledge to help you get the most out of Rust's features so that you can create robust and scalable programs. You'll begin with basics like creating functions, choosing data types, and binding variables, then move on to more advanced concepts. In addition to the countless code examples, you'll find three chapters dedicated to building complete projects: a number-guessing game, a Rust implementation of a command line tool, and a multithreaded server."--Publisher marketing},
+  isbn = {978-1-7185-0310-6},
+  lccn = {QA76.73.R87 K53 2023},
+  keywords = {Handbooks and manuals,Instructional and educational works,Rust (Computer program language)}
+}
+
 @article{knoxOpensourceDataManager2019,
   title = {An Open-Source Data Manager for Network Models},
   author = {Knox, Stephen and Tomlinson, James and Harou, Julien J. and Meier, Philipp and Rosenberg, David E. and Lund, Jay R. and Rheinheimer, David E.},
