### 2025-08-01 99ace5f
Add contributing.md link to readme
Related: https://github.com/Nadi-System/nadi-system/issues/5
