### 2025-08-01 95d9ab9
paper: Nadi -> NADI and add DSL's drawbacks
Related:
https://github.com/Nadi-System/nadi-system/issues/3
https://github.com/Nadi-System/nadi-system/issues/6
