+++ paper.md
@@ -32,10 +32,10 @@ metadata analysis as well as a plugin system to run user-defined
 functions on each node or the whole network. Plugins provide seamless
 integration with other softwares and programming languages.
 
-NADI can be used from the Command Line Interface (CLI), Graphical User
-Interface (GUI) using the NADI Integrated Development Environment
-(IDE), as a Rust library, or as a Python library, allowing users to
-write their own plugins or programs.
+NADI System can be used from the Command Line Interface (CLI),
+Graphical User Interface (GUI) using the NADI Integrated Development
+Environment (IDE), as a Rust library, or as a Python library, allowing
+users to write their own plugins or programs.
 
 # Statement of need
 Hydrological analysis sometimes consists of data that is related to
@@ -75,8 +75,9 @@ of the best ways to represent the river network
 [@rinaldoTreesNetworksHydrology2006;
 @kuhnDesigningLanguageSpatial2015;
 @abed-elmdoustEmergentSpectralProperties2017] --- and provide the DSL
-for network metadata analysis. NADI is written in Rust [@klabnikRustProgrammingLanguage2023] due to the memory
-safety [@fultonBenefitsDrawbacksAdopting2021;
+for network metadata analysis. Most components of the NADI System is
+written in Rust [@klabnikRustProgrammingLanguage2023] due to the
+memory safety [@fultonBenefitsDrawbacksAdopting2021;
 @xuMemorySafetyChallengeConsidered2021;
 @bugdenSafetyPerformanceProminent2022], runtime performances
 [@zhangUnderstandingRuntimePerformance2023], and the macro system that
@@ -88,7 +89,7 @@ The figure below shows the GUI of NADI IDE, with the editor (left top), function
 ![Screenshot of the NADI IDE showcasing different components](ide-screenshot.png)
 
 # Data Structures
-Nadi has the following main data structures:
+NADI System has the following main data structures:
 
 - **Node** is one point on the network. It can have input nodes, one output node, and attributes associated with it.
 - **Network** consists of several connected nodes. It can also have attributes associated with it.
@@ -96,7 +97,7 @@ Nadi has the following main data structures:
 - **Functions** are categorized into environment, network, and node functions based on what they work on. For example, network function is run on a network, while node function is run on each node.
 - **Expressions** are a combination of attributes, variables, function calls, conditionals, etc that can result in an attribute value.
 - **Propagation** is how the node functions are called in a network, you can call them in different order, based on a list, or filtered by expression.
-- **Task** in NADI is an execution body consisting of type of task, optional output attribute name, and expression or function call. Only the top level function call on the expression can be a mutable call.
+- **Task** in NADI System is an execution body consisting of type of task, optional output attribute name, and expression or function call. Only the top level function call on the expression can be a mutable call.
 - **Task Context** is the runtime environment for tasks to be run. It consists of a network, environmental variables, and functions loaded from plugins.
 
 Figure below shows how the different data types come together to generate a task and the task context. Each task runs in the task context, giving outputs, modifying the task context, or producing side effects (saving files).
@@ -117,10 +118,10 @@ $$
 
 Where, $x_i, y_i$ are values of $x,y$ on node $i$ and, $I_i$ is the set of input nodes for node $i$.
 
-@atreyaEstimatingInfluenceWater2024 demonstrates a complex task like river routing model using the network structure. For more concepts and up to date syntax, refer to the Nadi Book [@nadi-book-070].
+@atreyaEstimatingInfluenceWater2024 demonstrates a complex task like river routing model using the network structure. For more concepts and up to date syntax, refer to the NADI Book [@nadi-book-070].
 
 # Extensibility
[…]
