### 2025-08-01 Nadi-System/nadi-system#1 @Atreyagaurav
Thank you for catching this. I have now modified the compilation process so that any dead links are caught during that process. I was able to catch one more with this, and fix both of them. And I learned about using absolute link could result in problems while opening from file system directly, so that is also fixed now.
