### 2025-08-01 482ba61
paper: some grammar, clarity and references
Related: https://github.com/Nadi-System/nadi-system/issues/8
