+++ paper.md
@@ -26,63 +26,64 @@ bibliography: references.bib
 # Summary
 We present the Network Analysis and Data Integration (NADI) System, a
 developing software framework designed to facilitate river data
-analysis. NADI System comes with a Domain Specific Programming
-Language (DSL) that has a succinct and readable syntax for network
-metadata analysis as well as a plugin system to run user-defined
-functions on each node or the whole network. Plugins provide seamless
-integration with other softwares and programming languages.
+analysis. NADI System includes a Domain Specific Programming Language
+(DSL) that has a succinct and readable syntax for network metadata
+analysis as well as a plugin system to run user-defined functions on
+each node or the whole network. Plugins provide seamless integration
+with other softwares and programming languages.
 
 NADI System can be used from the Command Line Interface (CLI),
 Graphical User Interface (GUI) using the NADI Integrated Development
 Environment (IDE), as a Rust library, or as a Python library, allowing
-users to write their own plugins or programs.
+users to write their plugins or programs.
 
 # Statement of need
 Hydrological analysis sometimes consists of data that is related to
 points in the river, and frequently with relationships between
 upstream/downstream points (e.g., higher correlation, mass
-balance). Some analyses that benefit from use of such relationships
-are: finding inconsistencies in the data, filling missing data, and
-visualization of metadata. There is a need for intelligent
-computational assistance on network based system to reduce the
-workload that further improves the efficiency and reproducibility of
-research in this field [@rosenbergNextFrontierMaking2020]. Specific
-hydrology-focused softwares [@rossmanOpenSourcingEPANET2010;
+balance). Some analyses that benefit from the use of such
+relationships are: finding inconsistencies in the data, filling
+missing data, and visualization of metadata. There is a need for
+intelligent computational assistance on a network-based system to
+reduce the workload that further improves the efficiency and
+reproducibility of research in this field
+[@rosenbergNextFrontierMaking2020]. Specific hydrology-focused
+softwares [@rossmanOpenSourcingEPANET2010;
 @gironasNewApplicationsManual2010] lack general applicability, while
-general purpose programming languages might not have a succinct
+general-purpose programming languages might not have a succinct
 syntax. This highlights the need for a balanced approach that combines
 specificity to hydrological research questions with general
 capabilities.
 
-Domain Specific Programming Languages (DSLs) have several advantages
+Domain-Specific Programming Languages (DSLs) have several advantages
 including improved code readability and maintainability due to the
-tailored syntax and semantics, as well as increased efficiency through
-optimized algorithms and data structures
-[@mernikWhenHowDevelop2005]. Although Graphviz has made great progress
-on the visualization of graph [@gansnerOpenGraphVisualization2000;
-@ellsonGraphvizDynagraphStatic2004], it does not have the analytical
-capabilities. Hydrolang has the capabilities of analysis and
-visualization for hydrological applications, but it is web-based
-[@erazoramirezHydroLangOpensourceWebbased2022]. Languages made for
-grid-based spatial analysis [@pullarMapScriptMapAlgebra2001;
-@kuhnDesigningLanguageSpatial2015] are not suitable for values like
-streamflow which are partially spatially continuous along the river,
-but not in 2D/3D space.
+tailored syntax and semantics [@mernikWhenHowDevelop2005;
+albuquerqueQuantifyingUsabilityDomainspecific2015]. Among DSLs that
+have been developed for networks or hydrology, Graphviz focuses on
+graph visualization [@gansnerOpenGraphVisualization2000;
+@ellsonGraphvizDynagraphStatic2004], but lacks analytical
+capabilities. Hydrolang offers both analysis and visualization tools
[…]
