+++ references.bib
@@ -37,6 +37,23 @@
   keywords = {anastomosis,braiding,channel pattern,continuum concept,effective discharge curve,meandering}
 }
 
+@article{albuquerqueQuantifyingUsabilityDomainspecific2015,
+title = {Quantifying Usability of Domain-Specific Languages: {{An}} Empirical Study on Software Maintenance},
+shorttitle = {Quantifying Usability of Domain-Specific Languages},
+author = {Albuquerque, Diego and Cafeo, Bruno and Garcia, Alessandro and Barbosa, Simone and Abrah{\~a}o, Silvia and Ribeiro, Ant{\'o}nio},
+year = {2015},
+month = mar,
+journal = {Journal of Systems and Software},
+volume = {101},
+pages = {245--259},
+issn = {0164-1212},
+doi = {10.1016/j.jss.2014.11.051},
+url = {https://www.sciencedirect.com/science/article/pii/S0164121214002799},
+urldate = {2025-08-01},
+abstract = {A domain-specific language (DSL) aims to support software development by offering abstractions to a particular domain. It is expected that DSLs improve the maintainability of artifacts otherwise produced with general-purpose languages. However, the maintainability of the DSL artifacts and, hence, their adoption in mainstream development, is largely dependent on the usability of the language itself. Unfortunately, it is often hard to identify their usability strengths and weaknesses early, as there is no guidance on how to objectively reveal them. Usability is a multi-faceted quality characteristic, which is challenging to quantify beforehand by DSL stakeholders. There is even less support on how to quantitatively evaluate the usability of DSLs used in maintenance tasks. In this context, this paper reports a study to compare the usability of textual DSLs under the perspective of software maintenance. A usability measurement framework was developed based on the cognitive dimensions of notations. The framework was evaluated both qualitatively and quantitatively using two DSLs in the context of two evolving object-oriented systems. The results suggested that the proposed metrics were useful: (1) to early identify DSL usability limitations, (2) to reveal specific DSL features favoring maintenance tasks, and (3) to successfully analyze eight critical DSL usability dimensions.},
+keywords = {DSL,Metrics,Usability}
+}
+
 @article{arulGraphTheoryAlgorithms2023,
   title = {Graph {{Theory}} and {{Algorithms}} for {{Network Analysis}}},
   author = {Arul, Sharmila and Senthil, Gowri and Jayasudha, S. and Alkhayyat, Ahmed and Azam, Khalikov and Elangovan, R.},
