### 2025-08-01 Nadi-System/nadi-system#2 @Atreyagaurav
Thank you for this comment.

For Rust, I was able to find a satisfactory answer:
https://users.rust-lang.org/t/how-to-cite-the-rust-programming-language-in-academic-journals/123433

I have added the rust citation in the paper.md.

But python seems a bit problematic, do you have any suggestions on how I should cite python?

The official source says cite a favorite book, but I have never read any book for python. It doesn't feel right to cite a book I have never read.

https://docs.python.org/3/faq/general.html#are-there-any-published-articles-about-python-that-i-can-reference
