### 2025-08-01 Nadi-System/nadi-system#3 @Atreyagaurav
Thank you. I have a tendency to use Nadi as a proper noun, but because of the acronym it should be written "NADI".

For now, I have replaced all occurrences of Nadi into NADI. I have also made edits in documentation and paper to use NADI System while referring to the whole system of tools, while NADI as an adjective for components like: NADI GIS, NADI DSL, NADI Task System, etc.
