### 2025-08-01 Nadi-System/nadi-system#4 @Atreyagaurav
Thank you. I have fixed the text based on your suggestions.
