### 2025-08-01 Nadi-System/nadi-system#6 @Atreyagaurav
Thank you.

The world limit in the JOSS is only 1000, so I didn't go into details on a lot of things.

I added a sentence that talks about the tradeoffs in the Extensibility section. It also introduces the python library that can be used to load the NADI data structures and functions to use from Python. Which again has the tradeoff of not being able to use the syntax of NADI and other benefits it provides, but allows people to just code pure python while calling rust functions from NADI plugins if needed.
