### 2025-08-01 Nadi-System/nadi-system#8 @Atreyagaurav
Thank you

I'm a little confused about this. I didn't define mutable calls because I wanted to decrease word count. It is a programming term used in a paper about programming language after all. I have added a context now.

 As for Propagation and Task Context, they are defined in the Data Structures section. And that's also the first time they are mentioned. I have changed the sentences to hopefully improve the clarity.
