### 2025-08-02 Nadi-System/nadi-system#2 @amcandio
You can cite *Python 3 Reference Manual*. It doesn't imply you've read it cover to cover. It's a standard way to acknowledge the source of the software, much like citing a specification.
