### 2025-08-02 3af8810
Add citation for Python
Related: https://github.com/Nadi-System/nadi-system/issues/2
