### 2025-08-02 #8655 @amcandio
@Atreyagaurav I'm not seeing the latest changes in the pdf. For example, Python citation.
