### 2025-08-02 #8655 @Atreyagaurav
@amcandio it's in extensibility section. That's the first mention of python excluding the summary. Should I put it in the summary?
