### 2025-08-03 Nadi-System/nadi-system#9 @jhwohlgemuth
JOSS Review:  DSL = Domain Specific Language
"DSL" is more commonly known to represent "Domain Specific Language" vice "Domain Specific **_Programming_** Language".

Also, "Domain-specific Language" (DSL) is how it is used in a cited reference ("When and how to develop domain-specific languages"). Consistency would dictate using the same definition.

**Recommendation**: Use "Domain Specific Language" for DSL.

https://github.com/openjournals/joss-reviews/issues/8655
