### 2025-08-03 Nadi-System/nadi-system#10 @jhwohlgemuth
JOSS Review: Identify Linux build/test system dependencies
I was trying to run `cargo build --release` and `cargo test` on Debian and needed to install a handful of system dependencies. It would be useful for adoption and reproducibility if these dependencies were made explicit in `CONTRIBUTING.md` (or mdbook install instructions).

Required system dependencies:
	- `build-essential`
	- `libatk1.0-0`
	- `libglib2.0-dev`
	- `libgtk-3-dev`

https://github.com/openjournals/joss-reviews/issues/8655
