### 2025-08-03 Nadi-System/nadi-system#11 @jhwohlgemuth
JOSS Review: Security Considerations
This project relies on 7 dependencies identified as "unmaintained" by `cargo audit`

[`cargo-audit`](https://github.com/rustsec/rustsec) literally considers these "warnings", but you might consider alternatives (or at least be on the lookout for alternatives).

Integrating static analysis and security checks into the deployment pipeline (e.g., GitHub workflow) is also advised.

https://github.com/openjournals/joss-reviews/issues/8655
