### 2025-08-03 Nadi-System/nadi-system#12 @jhwohlgemuth
JOSS Review: Testing Improvements
**_Recommendations_**:
- Add code coverage creation to workflow using [`Tarpaulin`](https://github.com/xd009642/tarpaulin) and commit coverage report to the `nadi-system` repository.
  - I ran `cargo tarpaulin --out lcov --line --skip-clean` and found the coverage to be `23%`, which could be a little higher 🤓
- `nadi-serve` has unused imports
  ```rust
	- warning: unused imports: `ContentType` and `Header`
	 --> nadi-serve/src/main.rs:4:20
	  |
	4 | use rocket::http::{ContentType, Header};
	  |                    ^^^^^^^^^^^  ^^^^^^
	  |
	  = note: `#[warn(unused_imports)]` on by default
  ```

https://github.com/openjournals/joss-reviews/issues/8655
