### 2025-08-03 Nadi-System/nadi-system#13 @jhwohlgemuth
JOSS Review: Workspace Member Name Consistency
**_Recommendation_**:
> Make workspace member names all `nadi-name` or `nadi_name`

https://github.com/openjournals/joss-reviews/issues/8655
