### 2025-08-03 Nadi-System/nadi-system#14 @jhwohlgemuth
JOSS Review: Architecture Diagram
I understand JOSS papers have a strict length restrictions and that you cannot address everything, but considering the purpose of this paper I think it would be very useful to include an architecture diagram of `nadi-system`. If I had to choose, I would replace "figure 2" with an architecture diagram of the NADI system and its sub-modules.

**_Recommendation_**:
> Add architecture diagram to paper (replace figure 2?)

**_Fallback Recommendation_**:
> If you can't fit an architecture diagram into the paper, perhaps you could at least add an `ARCHITECTURE.md` ([inspiration](https://matklad.github.io/2021/02/06/ARCHITECTURE.md.html)) file to the project?

https://github.com/openjournals/joss-reviews/issues/8655
