### 2025-08-03 Nadi-System/nadi-system#15 @jhwohlgemuth
JOSS Review: DSL Description and Inspiration
I love DSLs. It would be awesome to have a better understanding of the NADI DSL after reading the paper. I understand DSL documentation is beyond the scope of this paper, but you could easily include a few words of comparison (e.g., "JSON superset", "XML-like", "Mermaid inspired", etc.). Additionally, more words on DSL design consideration would be applicable and appreciated.

In a related line of thinking, I would love to see a more extensive example of the NADI DSL in the paper.

Finally, I would love to see syntax highlighting extensions ([something along these lines](https://www.ethan-shea.com/posts/vs-code-embedded-language-syntax)) added to the roadmap (future work) 🤓

> **Note**: Not having syntax highlighting or editor support are two common tradeoffs regarding DSLs

https://github.com/openjournals/joss-reviews/issues/8655
