### 2025-08-04 5b7f4c0
Add details and some files for syntax highligh in other tech
Related: https://github.com/Nadi-System/nadi-system/issues/15
