+++ extra/syntax-highlight/README.md
@@ -1,0 +1,3 @@
+This contains syntax files for sublime (can be used with `typst`), web (using highlight.js), as well as examples for LaTeX and Typst.
+
+You can compile the LaTeX and typst file to see the results.
