### 2025-08-04 Nadi-System/nadi-system#15 @Atreyagaurav
Thank you very much!

The reason I did not put DSL examples in the paper is because the JOSS paper word limit is only 1000, which limits the content by a lot. I do have extensive examples in the documentation, is that not enough?  I could add some example, but I also wanted to avoid putting code examples in the paper in case some breaking syntax changes happens in the future.

I will add the syntax highlight extensions to the roadmap, it is there in my head, but not shown anywhere. I do have syntax highlight on the Web documentation using Highlight.js, in typst using sublime syntax (which should work with sublime and other editors, but I haven't tested it), and in NADI IDE using the internal parser.

Here is an example of syntax highlight using sublime syntax, you can see more examples in the PDF version of the Nadi Book (User Documentation):

 I need to figure out how to write `lsp` based server for nadi in the future to support multiple editors, but right now it is not a priority.
