### 2025-08-04 Nadi-System/nadi-system#15 @Atreyagaurav
And here is an example of syntax highlighting in the web version of the nadi book: https://nadi-system.github.io/example/ohio-validation.html

You can get the PDF version and look at the same chapter for PDF version of syntax highlight: https://nadi-system.github.io/data/nadi-book.pdf

I have added the codes/files used to generate those in the main repo now. I'm making them for documentation purposes only, I will probably get to making them for other editors after the main language is a bit established.
