### 2025-08-04 9304888
Add inspiration for DSL
Related: https://github.com/Nadi-System/nadi-system/issues/15
