### 2025-08-04 #8655 @amcandio
> @lolpro11 @amcandio - Thank you both for your timely feedback. I just want to confirm that the authors have addressed all your comments and that your review is now considered complete. Could you also please confirm your final recommendation regarding acceptance? Thanks!

Yes, the author has addressed all my feedback. I recommend the paper to be accepted.
