### 2025-08-04 fa3c431
Cargo fmt; add some more tests
Related: https://github.com/Nadi-System/nadi-system/issues/12
