+++ nadi_core/src/functions.rs
@@ -1,18 +1,19 @@
 #![allow(clippy::module_inception)]
 #![allow(non_local_definitions)] // warning from sabi_trait macro
 use crate::attrs::{AttrMap, AttrSlice};
-use crate::plugins::{NadiPlugin, load_library_safe};
+use crate::plugins::{load_library_safe, NadiPlugin};
 use crate::prelude::*;
-use crate::table::{ColumnAlign, contents_2_md};
+use crate::table::{contents_2_md, ColumnAlign};
 use abi_stable::std_types::Tuple2;
 use abi_stable::{
-    StableAbi, sabi_trait,
+    sabi_trait,
     std_types::{
+        map::REntry,
         RBox, RErr, RHashMap, ROk,
         ROption::{self, RNone, RSome},
         RResult, RString, RVec,
-        map::REntry,
     },
+    StableAbi,
 };
 use colored::Colorize;
 use std::collections::HashMap;
