### 2025-08-04 Nadi-System/nadi-system#12 @Atreyagaurav
Thank you for the review on the tests, as well as the tool for coverage testing. I didn't know about it.

I only have tests for nadi_core (by itself it has around 37% coverage), but a lot of functions associated with visual properties and standard output are hard to test. I am planning to add tests that runs the programming language and checks the stdout so that I can check error messages as well as output of the programs.

For now, I added some extra tests to improve it, the coverage around 41% on the nadi_core. I will investigate further into it this week.
