### 2025-08-04 Nadi-System/nadi-system#13 @Atreyagaurav
Thank you for the recommendation.

I already have `nadi-cli`, `nadi-py`, and `nadi_core` published in pypi/crates.io, so I don't know if I should/can change the names.
