### 2025-08-04 Nadi-System/nadi-system#9 @Atreyagaurav
Thank you!

I think I saw somewhere where they wrote "Domain Specific Programming Language (DSL)" and copied it (to distinguish it from markup/data languages). But what you said makes sense and I have changed it on the paper, I will change it in the documentation soon.
