### 2025-08-04 Nadi-System/nadi-system#10 @Atreyagaurav
Thank you for finding out the dependencies.

I have so far tested it on Arch Linux (my development env), Windows, Android (using termux) and MacOS. I never ran into this issue before, so I don't think we can say it is for all Linux.

I'll add an OS specific section later in the documentation after investigating further. I think looking into how GTK based projects have their instructions (since it looks like most of them come from it) might help me determine requirements for different OS (something like this: https://www.gtk.org/docs/installations/linux/#installing-gtk3-from-packages).
