### 2025-08-05 Nadi-System/nadi-system#11 @Atreyagaurav
Thank you for the review.

I ran the command and saw that the 7 errors comes from following dependencies:

| Error form | NADI Dependency|
|-|-|
| yaml-rust | iced |
|paste | iced/abi_stable|
|instant | iced |
| gtk-sys | rfd|
|generational-arena | abi_stable |
| gdk-sys | rfd |
| atk-sys | rfd|

I am using the dependencies for the following reasons:
- iced : GUI for the NADI IDE and other components (GTK was hard to compile/distribute for windows),
- abi_stable: Plugin system and data types because Rust doesn't have a stable ABI yet,
- rfd: cross-platform file browser

Among these, iced and abi_stable aren't something I can replace, and I don't think I should/could change their dependencies (recursively until these packages are replaced). I do have a fork of iced because of another bug, but because that's stopping me from publishing `nadi-ide` to crates.io, so I'm thinking of disregarding even that.

For rdf, I couldn't find any other solution that worked for all different OSes, but the problem here seems to be `sys` packages from gtk, so maybe I can find something to replace this package. I'll investigate further into this.
