### 2025-08-06 b8902c6
Add Task Architecture diagram showcasing the environment for DSL
