### 2025-08-06 Nadi-System/nadi-system#14 @Atreyagaurav
Thank you for the review.

I have replaced the figure 2 with a new diagram as shown below:

![New Figure 2](https://nadi-system.github.io/images/task-architecture.svg)

I have added the architecture.md on the repository (https://github.com/Nadi-System/nadi-system/commit/b8902c618ce1bc52d3ac4594ae352ed93fc05a74), as well as in [the documentation](https://nadi-system.github.io/devref/architecture.html#nadi-dsl)

I feel like the information provided by this diagram is less than the previous one, but I think this is better in giving a higher level overview of the System, and also less likely to be outdated in future versions.
