### 2025-08-06 625c444
Add OS specific instructions in README
Related: https://github.com/Nadi-System/nadi-system/issues/10
