### 2025-08-07 d3372bf
Add tests for nadi-cli and nadi_plugin
Related: https://github.com/Nadi-System/nadi-system/issues/12
