### 2025-08-09 5facb84
Add more tests; add funcs for shell env var; minimize highlight code
