### 2025-08-11 2df0dd7
test: Add more tests for nadi_plugin
