+++ nadi_plugin/src/lib.rs
@@ -6,10 +6,10 @@ use convert_case::{Case, Casing};
 use std::collections::HashMap;
 
 use proc_macro::TokenStream;
-use quote::{format_ident, quote, quote_spanned, ToTokens};
+use quote::{ToTokens, format_ident, quote, quote_spanned};
 use syn::{
-    parse_macro_input, punctuated::Punctuated, token::Comma, Attribute, DeriveInput, Expr, FnArg,
-    Ident, ItemFn, ItemMod, Lit, MetaNameValue, Type, TypeReference,
+    Attribute, DeriveInput, Expr, FnArg, Ident, ItemFn, ItemMod, Lit, MetaNameValue, Type,
+    TypeReference, parse_macro_input, punctuated::Punctuated, token::Comma,
 };
 
 #[derive(PartialEq)]
@@ -336,6 +336,10 @@ fn get_fn_arg(arg: &FnArg) -> (&Ident, &Type, FuncArgType) {
 }
 
 // HACK ignoring the path and assuming anything::Option is Option
+/// Checks if a type is Option
+///
+/// Any type with the name Option as the type name before the generics
+/// is considered as an option type.
 fn type_is_opt(ty: &Type) -> bool {
     // if let Type::Path(p) = ty {
     //     p.path.is_ident("Option")
@@ -390,7 +394,9 @@ fn get_opt_type(ty: &Type) -> Option<&Type> {
 /// external plugins compiled to cdynlib.
 #[proc_macro_attribute]
 pub fn nadi_plugin(args: TokenStream, item: TokenStream) -> TokenStream {
-    nadi_export_plugin(args, item, true)
+    let item = parse_macro_input!(item as ItemMod);
+    let args = parse_macro_input!(args with Punctuated<MetaNameValue, Comma>::parse_terminated);
+    nadi_export_plugin(args, item, true).into()
 }
 
 /// Register the `mod` as an internal plugin, only to be used on
@@ -398,12 +404,16 @@ pub fn nadi_plugin(args: TokenStream, item: TokenStream) -> TokenStream {
 /// top of the `mod` definition so it can see all the functions.
 #[proc_macro_attribute]
 pub fn nadi_internal_plugin(args: TokenStream, item: TokenStream) -> TokenStream {
-    nadi_export_plugin(args, item, false)
+    let item = parse_macro_input!(item as ItemMod);
+    let args = parse_macro_input!(args with Punctuated<MetaNameValue, Comma>::parse_terminated);
+    nadi_export_plugin(args, item, false).into()
 }
 
-fn nadi_export_plugin(_args: TokenStream, item: TokenStream, external: bool) -> TokenStream {
-    let item = parse_macro_input!(item as ItemMod);
-
+fn nadi_export_plugin(
+    _args: Punctuated<MetaNameValue, Comma>,
+    item: ItemMod,
+    external: bool,
+) -> proc_macro2::TokenStream {
     let name = &item.ident;
     let name_s = name.to_string();
     let name_mod = nadi_struct_name(name, "Mod");
@@ -506,7 +516,6 @@ fn nadi_export_plugin(_args: TokenStream, item: TokenStream, external: bool) -> 
             #item
         }
     }
-    .into()
 }
 
 fn get_nadi_functions<'a>(item: &'a ItemMod, funct: &'_ str) -> Vec<&'a Ident> {
@@ -919,7 +928,55 @@ fn format_docstrings(string: String) -> String {
 #[cfg(test)]
 mod test {
     use super::*;
+    use rstest::*;
     use syn::parse_quote;
+
+    #[rstest]
+    #[case(parse_quote!(Option<i32>), true)]
+    #[case(parse_quote!(i32), false)]
[…]
