### 2025-08-15 1b881ba
Added MaskedSeries support; code compiles; untested
