+++ nadi_core/src/timeseries.rs
@@ -1,9 +1,9 @@
-use crate::attrs::{type_name, Attribute, Date, DateTime, FromAttribute, Time};
+use crate::attrs::{Attribute, Date, DateTime, FromAttribute, Time, type_name};
 
 use abi_stable::{
-    external_types::RMutex,
-    std_types::{RArc, RHashMap, RString, RVec},
     StableAbi,
+    external_types::RMutex,
+    std_types::{RArc, RHashMap, ROption, RString, RVec},
 };
 
 pub type TimeLine = RArc<RMutex<TimeLineInner>>;
@@ -146,9 +146,17 @@ impl TimeSeries {
         &self.timeline
     }
 
-    pub fn values_as_attributes(&self) -> Vec<Attribute> {
-        self.values.clone().to_attributes()
+    pub fn len(&self) -> usize {
+        self.values.len()
     }
+
+    pub fn is_empty(&self) -> bool {
+        self.values.is_empty()
+    }
+
+    // pub fn values_as_attributes(&self) -> Vec<Attribute> {
+    //     self.values.clone().to_attributes()
+    // }
 
     pub fn series(&self) -> &Series {
         &self.values
@@ -156,6 +164,10 @@ impl TimeSeries {
 
     pub fn values<'a, T: FromSeries<'a>>(&'a self) -> Option<&'a [T]> {
         FromSeries::from_series(&self.values)
+    }
+
+    pub fn str_values<'a>(&'a self) -> Box<dyn Iterator<Item = String> + 'a> {
+        self.values.str_values()
     }
 
     pub fn values_mut<'a, T: FromSeries<'a>>(&'a mut self) -> Option<&'a mut [T]> {
@@ -187,6 +199,195 @@ impl TimeSeries {
 #[repr(C)]
 #[derive(StableAbi, Clone, PartialEq, Debug)]
 pub enum Series {
+    /// Masked Series and optional fill value
+    Masked(MaskedSeries, ROption<Attribute>),
+    /// Series without values
+    Complete(CompleteSeries),
+}
+
+/// Matches and calls respective functions for all variations of [`Series`]
+macro_rules! forward_funcs {
+    ($($func:ident -> $ret:ty),*) => {
+	impl Series {
+	    $(
+		pub fn $func(&self) -> $ret {
+		    match self {
+			Self::Masked(v, _) => v. $func (),
+			Self::Complete(v) => v. $func (),
+		    }
+		}
+	    )*
+	}
+    }
+}
+
+forward_funcs! {
+    len -> usize,
+    is_empty -> bool,
+    type_name -> &str
+}
+
+impl Series {
+    pub fn is_masked(&self) -> bool {
+        matches!(self, Self::Masked(_, _))
+    }
+
+    pub fn is_complete(&self) -> bool {
+        matches!(self, Self::Complete(_))
+    }
+
+    pub fn from_attr(vals: &Attribute, dtype: &str) -> Result<Self, String> {
+        CompleteSeries::from_attr(vals, dtype).map(Self::Complete)
+    }
+
+    pub fn str_values<'a>(&'a self) -> Box<dyn Iterator<Item = String> + 'a> {
+        match self {
+            Self::Masked(v, _) => v.str_values(),
+            Self::Complete(v) => v.str_values(),
+        }
+    }
+
+    pub fn maybe_complete(self) -> Self {
+        match self {
+            Self::Complete(_) => self,
+            Self::Masked(ms, fill) => {
+                if ms.has_gaps() {
+                    Self::Masked(ms, fill)
+                } else {
+                    Self::Complete(ms.complete().unwrap())
+                }
+            }
+        }
+    }
+}
+
[…]
