### 2025-08-16 Nadi-System/nadi-system#15 @Atreyagaurav
I tried it for vscode as well, the documentations on how to make one was a bit lacking (I don't know why `language-configuration.json` file made the grammar file not load at all).

I removed the configuration file so it doesn't have auto-pair of braces and such, but the highlight looks acceptable.

You can download and try the extension here: https://github.com/Nadi-System/nadi-task-system-vscode/releases/download/0.7.0/nadi-task-system-0.7.0.vsix

The extension code is hosted at: https://github.com/Nadi-System/nadi-task-system-vscode
