### 2025-08-17 001f57b
Add support functions for series; as well as plugin
