+++ nadi_core/src/parser/expressions.rs
@@ -13,6 +13,8 @@ use nom::{
     sequence::{delimited, pair, separated_pair, terminated, tuple},
 };
 
+// TODO: Add try-catch expression
+
 pub fn expression<'a, 'b>(inp: &'a [Token<'b>]) -> MatchRes<'a, 'b, Expression> {
     alt((
         input_variable,
