### 2025-08-18 1383af0
Add series support functions; add algo for gap filling by value
