+++ nadi_core/src/timeseries.rs
@@ -1,9 +1,9 @@
-use crate::attrs::{Attribute, Date, DateTime, FromAttribute, Time, type_name};
+use crate::attrs::{type_name, Attribute, Date, DateTime, FromAttribute, Time};
 
 use abi_stable::{
-    StableAbi,
     external_types::RMutex,
     std_types::{RArc, RHashMap, ROption, RSome, RString, RVec},
+    StableAbi,
 };
 
 pub type TimeLine = RArc<RMutex<TimeLineInner>>;
@@ -48,8 +48,22 @@ pub trait HasSeries {
             .get(name)
             .ok_or(format!("Series `{name}` not found"))
     }
-}
+    fn try_series_mut(&mut self, name: &str) -> Result<&mut Series, String> {
+        self.series_map_mut()
+            .get_mut(name)
+            .ok_or(format!("Series `{name}` not found"))
+    }
 
+    fn fill_series(&mut self, name: &str, value: Attribute) -> Result<(), String> {
+        let ser: Option<Series> = self.series_map_mut().remove(name).into();
+        self.set_series(
+            name,
+            ser.ok_or(format!("Series `{name}` not found"))?
+                .fill_gaps(value)?,
+        );
+        Ok(())
+    }
+}
 #[repr(C)]
 #[derive(StableAbi, Clone, Debug)]
 pub struct TimeLineInner {
@@ -266,6 +280,13 @@ impl Series {
             }
         }
     }
+
+    pub fn fill_gaps(self, value: Attribute) -> Result<Self, String> {
+        match self {
+            Self::Complete(_) => Ok(self),
+            Self::Masked(ms, _) => ms.fill_gaps(value).map(Self::Complete),
+        }
+    }
 }
 
 #[repr(C)]
@@ -416,6 +437,30 @@ impl MaskedSeries {
             Self::Attributes(v) => CompleteSeries::Attributes(to_complete(v)),
         })
     }
+
+    pub fn fill_gaps(self, value: Attribute) -> Result<CompleteSeries, String> {
+        Ok(match (self, value) {
+            (Self::Floats(v), Attribute::Float(f)) => CompleteSeries::Floats(fill_gaps(v, f)),
+            (Self::Floats(v), Attribute::Integer(i)) => {
+                CompleteSeries::Floats(fill_gaps(v, i as f64))
+            }
+            (Self::Integers(v), Attribute::Integer(f)) => CompleteSeries::Integers(fill_gaps(v, f)),
+            (Self::Strings(v), Attribute::String(f)) => CompleteSeries::Strings(fill_gaps(v, f)),
+            (Self::Booleans(v), Attribute::Bool(f)) => CompleteSeries::Booleans(fill_gaps(v, f)),
+            (Self::Dates(v), Attribute::Date(f)) => CompleteSeries::Dates(fill_gaps(v, f)),
+            (Self::Times(v), Attribute::Time(f)) => CompleteSeries::Times(fill_gaps(v, f)),
+            (Self::DateTimes(v), Attribute::DateTime(f)) => {
+                CompleteSeries::DateTimes(fill_gaps(v, f))
+            }
+            (Self::DateTimes(v), Attribute::Date(f)) => {
+                let f: DateTime = f.into();
+                CompleteSeries::DateTimes(fill_gaps(v, f))
+            }
+            (Self::Attributes(v), f) => CompleteSeries::Attributes(fill_gaps(v, f)),
+            _ => return Err("Data Type does not match with Series".to_string()),
+        })
+    }
+
     pub fn type_name(&self) -> &str {
         match self {
             Self::Floats(_) => "Floats",
