+++ nadi_core/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod attrs;
+pub mod calc;
 pub mod expressions;
 pub mod functions;
 pub mod network;
