### 2025-08-20 0c224b6
Add impl From<(T,...)> for Attribute
