### 2025-08-20 c910009
add function to read a single line from a file
