### 2025-08-26 Nadi-System/nadi-system#16 @mengqi-z
JOSS Paper Proofread
Hi @Atreyagaurav,

I have few comments for the paper openjournals/joss-reviews#8655.

1. Line 28 – Please fix the citation so that it is properly hyperlinked: `albuquerqueQuantifyingUsabilityDomainspecific2015`.
2. In `references.bib`, a large number of entries are not cited in `paper.bib`. Please review all references and remove any unused ones.

Please let me know once these are addressed. Thanks!
