### 2025-08-26 4ef7c83
Remove unused references in bibtex
Fixes: #16
