+++ references.bib
@@ -19,24 +19,6 @@
   keywords = {Geomorphology,Hydrology}
 }
 
-@article{alabyanTypesRiverChannel1998,
-  title = {Types of River Channel Patterns and Their Natural Controls},
-  author = {Alabyan, Andrei M. and Chalov, Roman S.},
-  year = {1998},
-  journal = {Earth Surface Processes and Landforms},
-  volume = {23},
-  number = {5},
-  pages = {467--474},
-  issn = {1096-9837},
-  doi = {10.1002/(SICI)1096-9837(199805)23:5<467::AID-ESP861>3.0.CO;2-T},
-  url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/%28SICI%291096-9837%28199805%2923%3A5%3C467%3A%3AAID-ESP861%3E3.0.CO%3B2-T},
-  urldate = {2024-12-03},
-  abstract = {River channel patterns are thought to form a morphological continuum. This continuum is two-dimensional, defined by plan features of which there are three (straight, meandering, branching), and structural levels of fluvial relief of which there are also three (floodplain, flood channel, low-water channel). Combinations of these three categories define the diversity of patterns. One of the most important factors in channel development is stream power, defined by water discharge and river slope. The greater the stream power, the stronger the branching tendency, but threshold values of stream power are different for the three different hierarchical levels of channel relief. The critical stream power values and hydrological regime together define the channel pattern, and analysis of the pattern type can be undertaken using effective discharge curves. {\copyright} 1998 John Wiley \& Sons, Ltd.},
-  copyright = {Copyright {\copyright} 1998 John Wiley \& Sons, Ltd.},
-  langid = {english},
-  keywords = {anastomosis,braiding,channel pattern,continuum concept,effective discharge curve,meandering}
-}
-
 @article{albuquerqueQuantifyingUsabilityDomainspecific2015,
   title = {Quantifying Usability of Domain-Specific Languages: {{An}} Empirical Study on Software Maintenance},
   shorttitle = {Quantifying Usability of Domain-Specific Languages},
@@ -54,33 +36,6 @@
   keywords = {DSL,Metrics,Usability}
 }
 
-@article{arulGraphTheoryAlgorithms2023,
-  title = {Graph {{Theory}} and {{Algorithms}} for {{Network Analysis}}},
-  author = {Arul, Sharmila and Senthil, Gowri and Jayasudha, S. and Alkhayyat, Ahmed and Azam, Khalikov and Elangovan, R.},
-  year = {2023},
-  month = jul,
-  journal = {E3S Web of Conferences},
-  volume = {399},
-  doi = {10.1051/e3sconf/202339908002},
-  abstract = {In network analysis, the study and comprehension of complex systems in numerous fields, such as social networks, transportation networks, and biological networks, are made possible by the crucial role played by graph theory and algorithms. In order to give a comprehensive review of the graph theory and network analysis methods, this abstract will focus on their significance, practical uses, and most recent developments. With items represented as nodes or vertices and links between them as edges, graph theory offers a mathematical framework for modeling and evaluating relationships between objects. Researchers may learn important things about the structure, connectivity, and behavior of complex systems by using graph theory in network analysis. As a result, network analysis is made possible by the graph theory and algorithms, which offer strong tools for studying and comprehending the complicated linkages and structures of complex systems. Graph theory and algorithms have many different applications, including social networks, transportation networks, and biological networks. Large-scale network analysis is now possible thanks to the development of effective algorithms and methodologies, which has significantly advanced the subject. The significance of graph theory and algorithms for network research will only rise as networks continue to expand in size and complexity.}
-}
-
[…]
