### 2025-08-26 #8655 @Atreyagaurav
I wonder if skipping the references not in paper, or flagging them in the `check references` bot command output, is a reasonable modification that can be useful for the bot? I do not have any experience in ruby, otherwise I could offer to add it.
