### 2025-08-27 486b91a
Add task Context in nadi-py; Update cargo version
