+++ nadi_core/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nadi_core"
-version = "0.7.1"
+version = "0.7.2"
 edition = "2021"
 description = "Core library for Nadi systems, for use by plugins"
 repository = "https://github.com/Nadi-System/nadi-system"
