### 2025-08-27 #8655 @Atreyagaurav
Hi @mengqi-z ,

I made the release for Version 0.7.2 that includes all changes during the review process.

- Zenodo: https://doi.org/10.5281/zenodo.16956958
- Github Release: https://github.com/Nadi-System/nadi-system/releases/tag/0.7.2

The repo for each language's libraries also now has 0.7.2 version.
- PyPi: https://pypi.org/project/nadi-py/
- Crates.io: https://docs.rs/nadi_core/latest/nadi_core/

I think everything looks good from my part on Author and affiliations.
