### 2025-08-27 0d7f35f
add ORCID for Patrick and Todd
