+++ paper.md
@@ -12,8 +12,10 @@ authors:
     affiliation: 1
   - name: Todd Steissberg
     affiliation: 2
+    orcid: 0000-0003-3744-701X
   - name: Patrick A. Ray
     affiliation: 1
+    orcid: 0000-0001-9495-2317
 affiliations:
   - name: Department of Chemical and Environmental Engineering, University of Cincinnati, OH, USA
     index: 1
