### 2025-08-27 #8655 @mengqi-z
@Atreyagaurav – I’ve recommended your submission for acceptance. Thanks so much for submitting to JOSS! From here, our Editor-in-Chief will step in to do some final checks before the paper can be officially accepted and published. Congrats and great job!

@jhwohlgemuth, @lolpro11, @amcandio – Huge thanks to all of you for reviewing this paper! I appreciate your timely feedback. I look forward to seeing your reviews for JOSS in the future!
