### 2025-08-27 #8655 @Atreyagaurav
Thank you very much for editing this submission. I'm grateful for the open and rapid review process of JOSS.
