### 2025-08-27 #8655 @Atreyagaurav
@kthyng Thank you very much! For the work on this paper as well as for having a journal like JOSS that focuses on being transparent, interactive, open access, and free. Overall it was a smooth process.
