### 2025-08-27 dbb3ef4
add citation.cff
