### 2025-08-27 #8655 @kthyng
Congratulations on your new publication @Atreyagaurav! Many thanks to editor @mengqi-z and to reviewers @jhwohlgemuth, @lolpro11, and @amcandio for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
